scripts/requirements.txt
psutil>=5.9.0
ruamel.yaml>=0.18.0
